+++ sites_faciles/blog/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class BlogConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "blog"
+    name ="sites_faciles.blog"
