+++ sites_faciles/content_manager/services/accessors.py
@@ -5,9 +5,9 @@ from wagtail.models import Collection, Page, PageViewRestriction
 from wagtailmenus.models.menuitems import MainMenuItem
 from wagtailmenus.models.menus import FlatMenu, MainMenu
 
-from sites_faciles.blog.constants import HEADER_FIELDS
-from sites_faciles.blog.models import CatalogIndexPage, ContentPage
-from sites_faciles.blog.utils import get_default_site
+from sites_faciles.content_manage.constants import HEADER_FIELDS
+from sites_faciles.content_manage.models import CatalogIndexPage, ContentPage
+from sites_faciles.content_manage.utils import get_default_site
 
 style = color_style()
